pyproject.toml
[tool.poetry]
name = "ovv-iso-chat"
version = "3.1.0"
description = "RAG System for ISO-SGI Documents with Long-Term Memory - Setup Minimo"
authors = ["OVV"]
readme = "docs/README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
# Core RAG
langgraph = "^0.2.0"
langchain = "^0.2.0"
langchain-core = "^0.2.0"
langchain-community = "^0.2.0"
langchain-ollama = "^0.1.0"

# Vector DB
qdrant-client = "^1.9.0"

# Embedding & Reranking
sentence-transformers = "^3.0.0"
FlagEmbedding = "^1.2.0"

# PDF Processing
pymupdf = "^1.23.0"
pdfplumber = "^0.10.0"

# Graph RAG (R25)
networkx = "^3.2"
community = "^1.0.0"  # python-louvain

# Utilities
python-dotenv = "^1.0.0"
pyyaml = "^6.0.0"
rich = "^13.0.0"
tqdm = "^4.66.0"
pydantic = "^2.0.0"

# PyTorch (CUDA)
torch = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"

[tool.poetry.group.ui]
optional = true

[tool.poetry.group.ui.dependencies]
gradio = "^4.0.0"
chainlit = "^1.0.0"
bcrypt = "^4.1.0"
streamlit = "^1.31.0"
plotly = "^5.18.0"

[tool.poetry.scripts]
ovv-chat = "src.main:main"
ovv-ingest = "src.ingestion.indexer:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

